feat: move certifications validation schemas to a new file for better organization
